requirements.txt
configparser
logging
requests
requests[security]
simplejson
GitPython
relayr
paho-mqtt
socketIO_client
certifi
